Move Block to its own file.
